move back_to_main_page_button to components folder
